@@ -39,11 +39,11 @@
     "@nuxtjs/storybook": "^4.1.1",
     "@nuxtjs/stylelint-module": "^4.0.0",
     "css-loader": "^5.2.7",
-    "eslint": "^7.30.0",
+    "eslint": "^7.31.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.23.4",
     "eslint-plugin-nuxt": "^2.0.0",
-    "eslint-plugin-vue": "^7.13.0",
+    "eslint-plugin-vue": "^7.14.0",
     "nuxt-use-motion": "^1.5.1",
     "nuxt-vite": "^0.1.1",
     "nuxt-windicss": "1.1.3",
